--- client/src/App.css
@@ -1,3 +1,0 @@
-body {
-  text-align: center;
-}
