+++ client/src/components/ToDoItems/ListItem.css
@@ -1,3 +1,8 @@
-.list.btn {
-  display: block;
+.list-item {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  padding: 0.5rem;
+  margin: 1rem 0;
+  background-color: grey;
 }
