+++ client/src/components/UserInput/UserInput.css
@@ -1,0 +1,46 @@
+.listuserinput {
+  width: 30rem;
+  max-width: 95%;
+  border-radius: 5px;
+  background-color: #f2f2f2;
+  padding: 20px;
+  margin: 0 auto 1rem;
+  display: flex;
+  flex-wrap: wrap;
+  gap: 1rem;
+  text-align: left;
+}
+
+.input-item label {
+  font-weight: bold;
+  margin-bottom: 0.5rem;
+  display: block;
+  text-align: left;
+}
+
+.input-item input {
+  font: inherit;
+  padding: 0.5rem;
+  border-radius: 6px;
+  border: 1px solid #ccc;
+  width: 20rem;
+  max-width: 100%;
+}
+
+.listuserinput button {
+  display: block;
+  margin: auto;
+  width: 100%;
+  background-color: #4caf50;
+  color: white;
+  padding: 14px 20px;
+  border: none;
+  border-radius: 4px;
+  cursor: pointer;
+}
+
+.listuserinput button:hover {
+  border: 1px solid black;
+  background-color: #fff;
+  color: #4caf50;
+}
