+++ client/src/pages/Login.js
@@ -1,49 +1,8 @@
-import React, { useState } from 'react';
+import React from 'react';
+import LoginForm from '../components/SignIn/LoginForm';
 
 const Login = (props) => {
-  const [email, setEmail] = useState('');
-  const [password, setPassword] = useState('');
-
-  const emailChangeHandler = (event) => {
-    setEmail(event.target.value);
-  };
-
-  const passwordChangeHandler = (event) => {
-    setPassword(event.target.value);
-  };
-
-  const handleSubmit = (event) => {
-    event.preventDefault();
-  };
-
-  return (
-    <form onSubmit={handleSubmit}>
-      <label htmlFor="email">Email</label>
-      <input
-        type="email"
-        name="email"
-        placeholder="example@email.com"
-        id="email"
-        required
-        onChange={emailChangeHandler}
-        value={email}
-      />
-      <label htmlFor="password">Password</label>
-      <input
-        type="password"
-        name="password"
-        placeholder="password"
-        id="password"
-        required
-        onChange={passwordChangeHandler}
-        value={password}
-      />
-      <button type="submit">Login</button>
-      <button>
-        <Link to="/register">Don't have an account</Link>
-      </button>
-    </form>
-  );
+  return <LoginForm />;
 };
 
 export default Login;
