+++ client/src/pages/ToDoList.js
@@ -1,6 +1,6 @@
 import React, { useCallback, useEffect, useState } from 'react';
 import ListResults from '../components/ToDoItems/ListResults';
-import UserInput from '../components/ToDoItems/UserInput';
+import UserInput from '../components/UserInput/UserInput';
 
 const UserData = () => {
   const [tasks, setTasks] = useState([]);
